Recover fn_8014E1DC (s16 lvalue compound-add extends sum, not RHS cast)
The face-the-player turn `*(s16*)obj += ...` cast the RHS to (s16),
emitting extsh on the converted delta before the add. The retail target
extends the *sum* after `add r3,r0` (the s16 lvalue's own narrowing),
so drop the explicit (s16) cast. 98.16 -> 98.74.

+++ src/main/dll/dll_00DF_hagabon.c
@@ -273,7 +273,7 @@ void fn_8014E1DC(int obj, HagabonState* state)
         angleDelta += 0xffff;
     }
 
-    *(s16*)obj += (s16)(s32)(((f32)angleDelta * timeDelta) / lbl_803E263C);
+    *(s16*)obj += (s32)(((f32)angleDelta * timeDelta) / lbl_803E263C);
 }
 
 void hagabon_hitDetect(int obj)
